Refactor bootloader code, memory map

+++ Makefile
@@ -3,17 +3,14 @@
 source := src
 target := target
 
-all: clean create-target build-all
-build-all: build-bootloader
-build-bootloader: compile-boot0 compile-boot1 merge-boot
+all: clean create-target compile-boot
 
 clean:
-	rmdir "${target}\\" /S /Q
+	if exist "${target}\\" rmdir "${target}\\" /S /Q
 create-target:
 	if not exist "${target}\\" mkdir "${target}\\"
-compile-boot0:
+compile-boot:
+	nasm ${source}\boot\glob.asm -f bin -o ${target}\glob.bin
 	nasm ${source}\boot\boot0.asm -f bin -o ${target}\boot0.bin
-compile-boot1:
 	nasm ${source}\boot\boot1.asm -f bin -o ${target}\boot1.bin
-merge-boot:
-	copy ${target}\boot0.bin /B + ${target}\boot1.bin /B ${target}\os.bin /B
+	copy ${target}\boot0.bin /B + ${target}\boot1.bin /B + ${target}\glob.bin /B ${target}\os.bin /B
